trading-bot/requirements.txt
ccxt>=4.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
requests>=2.31.0
pandas_ta>=0.4.7